Add Title command feature
This feature depends on the removal of the prompt escape sequences that
came before. Part of that is the removal of adding the title to $PS1:
now it is just print to the screen. This could improve speed slightly,
as the shell doesn't need to parse that text. But it does prevent usage
of prompt escape codes in the title.

I also needed to rework how the runtime feature was handled, since it
monopolized the DEBUG trap in Bash. Now it can handle multiple features
that need to run after the user has submitted the command, but before it
runs.

The feature itself is simple: lookup the latest command in the history,
and print the normal title postpended by that command to the screen,
formatted as a terminal title. Note that Bash's $BASH_COMMAND is not
great, as it only contains the current _command_, not the whole command
line.

Resolves #609